pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
authors = ["Sergeev O. V. <o.v.sergeyev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
pydantic = "^2.10.3"
pydantic-settings = "^2.6.1"
pyjwt = "^2.10.1"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
